requirements.txt
Flask
psycopg2-binary
pandas
openpyxl
pyxlsb